# GH-3941 improve iterations