--- a/core/common/iterator/src/main/java/org/eclipse/rdf4j/common/iteration/UnionIteration.java
+++ b/core/common/iterator/src/main/java/org/eclipse/rdf4j/common/iteration/UnionIteration.java
@@ -38,6 +38,7 @@ public class UnionIteration<E, X extends Exception> extends LookAheadIteration<E
 	 *
 	 * @param args The Iterations containing the elements to iterate over.
 	 */
+	@SafeVarargs
 	public UnionIteration(Iteration<? extends E, X>... args) {
 		this(Arrays.asList(args));
 	}