--- a/core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/QueryEvaluationStep.java
+++ b/core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/QueryEvaluationStep.java
@@ -45,6 +45,9 @@ public interface QueryEvaluationStep {
 		}
 	}
 
+	EmptyIteration<BindingSet, QueryEvaluationException> EMPTY_ITERATION = new EmptyIteration<>();
+	QueryEvaluationStep EMPTY = bindings -> EMPTY_ITERATION;
+
 	CloseableIteration<BindingSet, QueryEvaluationException> evaluate(BindingSet bindings);
 
 	/**