--- a/core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/impl/ExtensionQueryEvaluationStep.java
+++ b/core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/impl/ExtensionQueryEvaluationStep.java
@@ -3,7 +3,6 @@ package org.eclipse.rdf4j.query.algebra.evaluation.impl;
 import java.util.function.Consumer;
 
 import org.eclipse.rdf4j.common.iteration.CloseableIteration;
-import org.eclipse.rdf4j.common.iteration.EmptyIteration;
 import org.eclipse.rdf4j.query.BindingSet;
 import org.eclipse.rdf4j.query.MutableBindingSet;
 import org.eclipse.rdf4j.query.QueryEvaluationException;
@@ -32,7 +31,7 @@ public final class ExtensionQueryEvaluationStep implements QueryEvaluationStep {
 			// a type error in an extension argument should be silently ignored
 			// and
 			// result in zero bindings.
-			result = new EmptyIteration<>();
+			result = QueryEvaluationStep.EMPTY_ITERATION;
 		}
 		return new ExtensionIterator(result, consumer, context);
 	}