--- a/core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/impl/TupleFunctionEvaluationStrategy.java
+++ b/core/queryalgebra/evaluation/src/main/java/org/eclipse/rdf4j/query/algebra/evaluation/impl/TupleFunctionEvaluationStrategy.java
@@ -123,9 +123,9 @@ public class TupleFunctionEvaluationStrategy extends StrictEvaluationStrategy {
 	public static CloseableIteration<BindingSet, QueryEvaluationException> evaluate(TupleFunction func,
 			final List<Var> resultVars, final BindingSet bindings, ValueFactory valueFactory, Value... argValues)
 			throws QueryEvaluationException {
-		final CloseableIteration<? extends List<? extends Value>, QueryEvaluationException> iter = func
-				.evaluate(valueFactory, argValues);
 		return new LookAheadIteration<>() {
+			private final CloseableIteration<? extends List<? extends Value>, QueryEvaluationException> iter = func
+					.evaluate(valueFactory, argValues);
 
 			@Override
 			public BindingSet getNextElement() throws QueryEvaluationException {
@@ -158,11 +158,7 @@ public class TupleFunctionEvaluationStrategy extends StrictEvaluationStrategy {
 
 			@Override
 			protected void handleClose() throws QueryEvaluationException {
-				try {
-					super.handleClose();
-				} finally {
-					iter.close();
-				}
+				iter.close();
 			}
 		};
 	}