GitHub issue resolved: #3941 <!-- add a Github issue number here, e.g #123. -->

Briefly describe the changes proposed in this PR:

<!-- short description of your change goes here -->


## Before
```
Benchmark                                                     Mode  Cnt      Score     Error  Units
QueryBenchmark.complexQuery                                   avgt    5      2.987 ±   0.264  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5      1.401 ±   0.041  ms/op
QueryBenchmark.groupByQuery                                   avgt    5      0.994 ±   0.049  ms/op
QueryBenchmark.long_chain                                     avgt    5    521.931 ±  16.653  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    121.140 ±   4.662  ms/op
QueryBenchmark.minus                                          avgt    5  14466.794 ± 637.825  ms/op
QueryBenchmark.nested_optionals                               avgt    5    131.539 ±   4.769  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     13.273 ±   0.896  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5      1.230 ±   0.032  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5     79.481 ±   3.536  ms/op
QueryBenchmark.simple_filter_not                              avgt    5      5.504 ±   0.805  ms/op
```

```
Benchmark                                    Mode  Cnt   Score   Error  Units
ParallelQueryBenchmark.mixedQueriesAndReads  avgt    5  55.055 ± 1.288  ms/op
```

```
Benchmark                                         Mode  Cnt   Score   Error  Units
SparqlOverheadBenchmark.getAllStatementSparql     avgt    5  34.649 ± 2.038  ms/op
SparqlOverheadBenchmark.getAllStatementsDirectly  avgt    5  12.285 ± 0.314  ms/op
SparqlOverheadBenchmark.queryWithManyVars         avgt    5   0.329 ± 0.002  ms/op
```

## After
```
Benchmark                                                     Mode  Cnt     Score    Error  Units
QueryBenchmark.complexQuery                                   avgt    5     2.656 ±  0.075  ms/op
QueryBenchmark.different_datasets_with_similar_distributions  avgt    5     1.199 ±  0.011  ms/op
QueryBenchmark.groupByQuery                                   avgt    5     0.927 ±  0.107  ms/op
QueryBenchmark.long_chain                                     avgt    5   457.816 ±  6.422  ms/op
QueryBenchmark.lots_of_optional                               avgt    5    97.436 ±  0.898  ms/op
QueryBenchmark.minus                                          avgt    5  2450.321 ± 38.688  ms/op
QueryBenchmark.nested_optionals                               avgt    5   108.491 ±  1.954  ms/op
QueryBenchmark.pathExpressionQuery1                           avgt    5     9.069 ±  0.466  ms/op
QueryBenchmark.pathExpressionQuery2                           avgt    5     1.073 ±  0.038  ms/op
QueryBenchmark.query_distinct_predicates                      avgt    5    75.272 ±  1.035  ms/op
QueryBenchmark.simple_filter_not                              avgt    5     4.186 ±  0.036  ms/op
```

```
Benchmark                                    Mode  Cnt   Score   Error  Units
ParallelQueryBenchmark.mixedQueriesAndReads  avgt    5  35.535 ± 0.635  ms/op
```

```
Benchmark                                         Mode  Cnt   Score   Error  Units
SparqlOverheadBenchmark.getAllStatementSparql     avgt    5  29.754 ± 2.379  ms/op
SparqlOverheadBenchmark.getAllStatementsDirectly  avgt    5  12.316 ± 0.488  ms/op
SparqlOverheadBenchmark.queryWithManyVars         avgt    5   0.292 ± 0.015  ms/op
```

PR Author Checklist (see the [contributor guidelines](https://github.com/eclipse/rdf4j/blob/main/.github/CONTRIBUTING.md) for more details):

 - [ ] my pull request is [self-contained](https://rdf4j.org/documentation/developer/merge-strategy/#self-contained-changes-pull-requests-and-commits)
 - [ ] I've added tests for the changes I made
 - [ ] I've applied [code formatting](https://github.com/eclipse/rdf4j/blob/main/.github/CONTRIBUTING.md#code-formatting) (you can use `mvn process-resources` to format from the command line)
 - [ ] I've [squashed](https://rdf4j.org/documentation/developer/squashing) my commits where necessary 
 - [ ] every commit message starts with the issue number (GH-xxxx) followed by a meaningful description of the change